@@ -1,74 +1,65 @@
 {
-  "name": "ngx-tweet",
-  "version": "1.0.1",
-  "description": "Angular component for displaying a Twitter Tweet",
-  "author": "Raphaël MARQUES <work.rmarques@gmail.com> (https://raphaelmarques.fr)",
-  "homepage": "https://github.com/Adrael/ngx-tweet#readme",
-  "license": "MIT",
-  "main": "./src/index.ts",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "tslint \"src/**/*.ts\"",
-    "pree2e": "webdriver-manager update",
-    "e2e": "protractor",
-    "tsc": "tsc",
-    "tsc:w": "tsc -w",
-    "prepublish": "tsc -p ./tsconfig.publish.json",
-    "clean:aot": "rimraf aot/*",
-    "clean:lib": "rimraf lib/*",
-    "clean:lib-temp": "rimraf lib-temp/*",
-    "clean": "npm run clean:aot && npm run clean:lib && npm run clean:lib-temp",
-    "copy:assets": "cpx \"src/assets/**/*\" lib/assets",
-    "copy:src": "cpx \"lib-temp/src/**/*\" lib",
-    "copy": "npm run copy:assets && npm run copy:src",
-    "ngc": "node_modules/.bin/ngc -p tsconfig-aot.json",
-    "b": "npm run clean && npm run ngc && npm run copy && npm run clean:lib-temp"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Adrael/ngx-tweet.git"
-  },
-  "keywords": [
-    "Angular",
-    "Twitter",
-    "Tweet"
-  ],
-  "bugs": {
-    "url": "https://github.com/Adrael/ngx-tweet/issues"
-  },
-  "typings": "./lib/index.d.ts",
-  "dependencies": {},
-  "devDependencies": {
-    "@angular/common": "^5.0.0",
-    "@angular/compiler": "^5.0.0",
-    "@angular/compiler-cli": "^5.0.0",
-    "@angular/core": "^5.0.0",
-    "@angular/forms": "^5.0.0",
-    "@angular/http": "^5.0.0",
-    "@angular/platform-browser": "^5.0.0",
-    "@angular/platform-browser-dynamic": "^5.0.0",
-    "@angular/platform-server": "^5.0.0",
-    "@angular/router": "^5.0.0",
-    "core-js": "^2.5.1",
-    "rxjs": "^5.5.2",
-    "zone.js": "^0.8.18",
-    "ts-helpers": "^1.1.2",
-    "@types/chai": "^4.0.4",
-    "@types/jasmine": "^2.6.3",
-    "@types/node": "^8.0.50",
-    "codelyzer": "4.0.1",
-    "cpx": "^1.5.0",
-    "jasmine-core": "^2.8.0",
-    "jasmine-spec-reporter": "^4.2.1",
-    "karma": "^1.7.1",
-    "karma-chrome-launcher": "^2.2.0",
-    "karma-cli": "^1.0.1",
-    "karma-jasmine": "^1.1.0",
-    "karma-remap-istanbul": "^0.6.0",
-    "protractor": "^5.2.0",
-    "rimraf": "^2.6.2",
-    "ts-node": "^3.3.0",
-    "tslint": "^5.8.0",
-    "typescript": "^2.6.1"
-  }
+    "name": "ngx-tweet",
+    "version": "2.0.0",
+    "description": "Angular component for displaying a Twitter Tweet",
+    "author": "Raphaël MARQUES <work.rmarques@gmail.com> (https://raphaelmarques.fr)",
+    "homepage": "https://github.com/Adrael/ngx-tweet#readme",
+    "license": "MIT",
+    "private": false,
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "build:prod": "ng build --prod",
+        "test": "ng test",
+        "lint": "ng lint",
+        "e2e": "ng e2e",
+        "packagr": "ng-packagr -p ng-package.json"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Adrael/ngx-tweet.git"
+    },
+    "keywords": [
+        "Angular",
+        "Twitter",
+        "Tweet"
+    ],
+    "bugs": {
+        "url": "https://github.com/Adrael/ngx-tweet/issues"
+    },
+    "devDependencies": {
+        "@angular/animations": "6.0.2",
+        "@angular/common": "6.0.2",
+        "@angular/compiler": "6.0.2",
+        "@angular/core": "6.0.2",
+        "@angular/forms": "6.0.2",
+        "@angular/http": "6.0.2",
+        "@angular/platform-browser": "6.0.2",
+        "@angular/platform-browser-dynamic": "6.0.2",
+        "@angular/router": "6.0.2",
+        "core-js": "2.5.4",
+        "rxjs": "6.0.0",
+        "zone.js": "0.8.26",
+        "@angular-devkit/build-angular": "0.6.3",
+        "@angular/cli": "6.0.3",
+        "@angular/compiler-cli": "6.0.2",
+        "@angular/language-service": "6.0.2",
+        "@types/jasmine": "2.8.6",
+        "@types/jasminewd2": "2.0.3",
+        "@types/node": "8.9.4",
+        "codelyzer": "4.2.1",
+        "jasmine-core": "2.99.1",
+        "jasmine-spec-reporter": "4.2.1",
+        "karma": "1.7.1",
+        "karma-chrome-launcher": "2.2.0",
+        "karma-coverage-istanbul-reporter": "1.4.2",
+        "karma-jasmine": "1.1.1",
+        "karma-jasmine-html-reporter": "0.2.2",
+        "ng-packagr": "3.0.0",
+        "protractor": "5.3.0",
+        "ts-node": "5.0.1",
+        "tslint": "5.9.1",
+        "typescript": "2.7.2"
+    }
 }
